import-graphql-files-extract-queries * Move useNearestPlaces to its own directory * Extract graphql query to its own file and add @api directive * Change from using old apolloClient to new
