+++ src/logic/use-stop-places-with-departures/useStopPlacesWithDepartures.ts
@@ -5,7 +5,7 @@ import { LineData, StopPlaceWithDepartures } from '../../types'
 import { isNotNullOrUndefined, nonEmpty, unique } from '../../utils'
 import { useSettings } from '../../settings/SettingsProvider'
 import { REFRESH_INTERVAL } from '../../constants'
-import { useNearestPlaces } from '../useNearestPlaces'
+import { useNearestPlaces } from '../use-nearest-places/useNearestPlaces'
 import { apolloClient } from '../../services/realtimeVehicles/realtimeVehiclesService'
 import GET_STOP_PLACES_WITH_DEPARTURES_QUERY from './GetStopPlacesWithDepartures.graphql'
